+++ src/service.ts
@@ -2,11 +2,12 @@
 /**
  * service module
  */
+import { service } from '@chevre/domain';
+
 import * as AccountService from './service/account';
 import * as CodeService from './service/code';
 import * as CustomerService from './service/customer';
 import * as DeliveryService from './service/delivery';
-import * as IAMService from './service/iam';
 import * as NotificationService from './service/notification';
 import * as OfferService from './service/offer';
 import * as OrderService from './service/order';
@@ -15,7 +16,6 @@ import * as ProductService from './service/product';
 import * as ProjectService from './service/project';
 import * as ReportService from './service/report';
 import * as ReservationService from './service/reservation';
-import * as TaskService from './service/task';
 import * as TransactionService from './service/transaction';
 import * as UtilService from './service/util';
 
@@ -23,17 +23,15 @@ export import account = AccountService;
 export import code = CodeService;
 export import customer = CustomerService;
 export import delivery = DeliveryService;
-export import iam = IAMService;
+export import iam = service.iam;
 export import notification = NotificationService;
 export import offer = OfferService;
 export import order = OrderService;
-export namespace person {
-}
 export import product = ProductService;
 export import report = ReportService;
 export import reservation = ReservationService;
 export import payment = PaymentService;
 export import project = ProjectService;
-export import task = TaskService;
+export import task = service.task;
 export import transaction = TransactionService;
 export import util = UtilService;