+++ CHANGELOG.md
@@ -45,6 +45,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - 資産取引を再定義したchevreで再構築
 - 全chevreサービスにプロジェクトを指定
 - @chevre/factoryで再構築
+- Mongoリポジトリを@chevre/domainからインポートするように変更
 
 ### Deprecated